Renamed Log class
