+++ library/src/main/java/com/wultra/android/powerauth/networking/log/WPNLogger.kt
@@ -1,22 +1,7 @@
-/*
- * Copyright 2022 Wultra s.r.o.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions
- * and limitations under the License.
- */
-
-package com.wultra.android.powerauth.networking
+package com.wultra.android.powerauth.networking.log
 
 import android.util.Log
+import com.wultra.android.powerauth.networking.ECIESInterceptor
 import okhttp3.Headers
 import okhttp3.Interceptor
 import okhttp3.OkHttpClient
@@ -30,7 +15,7 @@ import java.net.URL
  *
  * Logs are written with "WPN" tag to standard [android.util.Log] logger.
  */
-class Logger {
+class WPNLogger {
 
     enum class VerboseLevel {
         /** Silences all messages. */
@@ -165,7 +150,6 @@ class Logger {
         }
     }
 }
-
 
 private val headersToSkp = listOf(
     "accept-language", "content-type", "content-length", "accept-language", "transfer-encoding", "date", "server", "user-agent",
